log2incident/etl/flink_job.py
from pyflink.datastream import StreamExecutionEnvironment
from pyflink.table import StreamTableEnvironment
from log2incident.models import TaggedLog
import json

def flink_filter():
    env = StreamExecutionEnvironment.get_execution_environment()
    t_env = StreamTableEnvironment.create(env)

    # Assume data from S3 or stream
    # For simplicity, placeholder

    # Filter logs with 'error' tag
    # This is a basic example; in real, connect to S3 stream or Kafka

    # Placeholder: process a list of logs
    logs = []  # In real, read from source

    filtered_logs = [log for log in logs if 'error' in log.tags]

    # Trigger events for filtered logs
    for log in filtered_logs:
        # Call event creator
        pass

    env.execute("Log Filter Job")