pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "log2incident"
version = "0.1.0"
description = "Log processing pipeline from raw logs to incidents"
authors = [{name = "Your Name", email = "your.email@example.com"}]
dependencies = [
    "boto3>=1.34.0",
    "apache-flink>=1.18.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "moto>=4.0.0"
]
requires-python = ">=3.8"

[tool.setuptools.packages.find]
where = ["."]
include = ["log2incident*"]